requirements.txt
langchain-openai==0.0.2
langchain==0.1.0
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
faiss-cpu==1.7.4
numpy>=1.22.0
tiktoken==0.5.2
openai==0.27.8 